Refactor backend lambda function
Splits the route handlers into their own file. We don't count
express-server.js in the test coverage; express-route-handlers.js is the
file we care about coverage on. Also exposes functions for testing by
exporting them.
